Add tests for counter
