+++ package.json
@@ -26,6 +26,7 @@
     "@babel/cli": "^7.21.0",
     "@babel/core": "^7.21.3",
     "@babel/preset-env": "^7.20.2",
+    "@testing-library/jest-dom": "^5.16.5",
     "babel-eslint": "^10.1.0",
     "babel-preset-env": "^1.7.0",
     "css-loader": "^6.7.3",
